minmax

+++ src/UCTSearch.h
@@ -114,7 +114,7 @@ public:
 	
 	SearchResult play_simulation(GameState& currstate, UCTNode* const node);
 	SearchResult play_simulation_sh(GameState& currstate, UCTNode* const node);
-	int shot(GameState& currstate, UCTNode* node, Random& rd, int buget, int& budgetUsed, int& playouts, double& wins, bool isroot, int po_res_mode, int sim_res_mode,int bestmove);
+	double shot(GameState& currstate, UCTNode* node, Random& rd, int buget, int& budgetUsed, int& playouts, double& wins, bool isroot, int po_res_mode, int sim_res_mode,int bestmove);
 private:
     float get_min_psa_ratio() const;
     void dump_stats(FastState& state, UCTNode& parent);
